Add explicit perl-interpreter buildrequires (effective no-op)
This is effectively a no-op, as with EL 6 & 7 (likely not 8)
perl is included in the buildsystem building group, and
with Fedora other buildrequires pulls in perl-interpreter,
however the Fedora perl packaging guidelines explicitly state
that "perl-interpreter must be listed as a build dependency
if it is called in any way".  The word "must" typically has a
very specific interpretation in technical docs, so follow the
"must".